Move FilterExemptions inside AdminFilters
