--- admin/app/conf/AdminFilters.scala
@@ -1,16 +1,0 @@
-package conf
-
-import com.gu.googleauth.FilterExemption
-
-object FilterExemptions {
-
-  lazy val loginExemption: FilterExemption = FilterExemption("/login")
-  lazy val exemptions: Seq[FilterExemption] = List(
-    FilterExemption("/oauth2callback"),
-    FilterExemption("/assets"),
-    FilterExemption("/_healthcheck"),
-    FilterExemption("/deploys-notify") // This endpoint is authenticated via api-key (to be accessible to riffraff hook for instance)
-  )
-}
-
-
